Put a new option called Enchanced Rotation Speed.
What this does is remove the 300ms delay after ever AA/Spell/Item. "
If this is disabled, it will make everything go quite fast, however you run into situations where Buffs/Debuffs and your /CheckFors will not work quite correctly, as the server has not been given enough time for your client to get updated information on the buff. A previous release removed this without me noticing that issue so I have rolled it back to a default of off, where you can turn it back on for faster processing.